Expand ZwiftEvent models, update subgroupId to long
Expanded ZwiftEventResponse and ZwiftEventSubgroupResponse with new required properties for event details, counts, tags, and more. Changed Id and related fields from int to long. Updated ZwiftHttpClient methods to use long for eventSubgroupId. Minor whitespace cleanup in ZwiftGameInfoService.

+++ src/FitSyncHub.Zwift/HttpClients/ZwiftHttpClient_Events.cs
@@ -129,7 +129,7 @@ public sealed partial class ZwiftHttpClient
     }
 
     public async Task<ZwiftRaceResultResponse> GetEventSubgroupResults(
-        int eventSubgroupId,
+        long eventSubgroupId,
         CancellationToken cancellationToken)
     {
         List<ZwiftRaceResultEntryResponse> acc = [];
@@ -164,7 +164,7 @@ public sealed partial class ZwiftHttpClient
     }
 
     public async Task<IReadOnlyCollection<ZwiftEventSubgroupEntrantResponse>> GetEventSubgroupEntrants(
-        int eventSubgroupId,
+        long eventSubgroupId,
         string type = "all", // or 'leader', 'sweeper', 'favorite', 'following', 'other'
         string participation = "signed_up", // or 'registered',
         CancellationToken cancellationToken = default)
